+++ notificare-loyalty/src/main/java/re/notifica/loyalty/internal/network/push/Payloads.kt
@@ -1,9 +1,8 @@
 package re.notifica.loyalty.internal.network.push
 
 import com.squareup.moshi.JsonClass
-import re.notifica.models.NotificareTransport
 
 @JsonClass(generateAdapter = true)
 internal data class RegisterPassPayload(
-    val transport: NotificareTransport,
+    val transport: String,
 )