+++ notificare-push/src/main/java/re/notifica/push/internal/ServiceManager.kt
@@ -2,7 +2,7 @@ package re.notifica.push.internal
 
 import re.notifica.InternalNotificareApi
 import re.notifica.internal.AbstractServiceManager
-import re.notifica.models.NotificareTransport
+import re.notifica.push.models.NotificareTransport
 
 @InternalNotificareApi
 public abstract class ServiceManager : AbstractServiceManager() {