+++ notificare/src/main/java/re/notifica/internal/modules/NotificareDeviceModuleImpl.kt
@@ -2,103 +2,104 @@ package re.notifica.internal.modules
 
 import android.content.Intent
 import androidx.annotation.Keep
-import java.util.Calendar
-import java.util.Date
-import java.util.GregorianCalendar
-import java.util.Locale
-import java.util.UUID
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.launch
 import kotlinx.coroutines.withContext
 import re.notifica.Notificare
 import re.notifica.NotificareCallback
 import re.notifica.NotificareDeviceModule
-import re.notifica.NotificareInternalDeviceModule
+import re.notifica.NotificareDeviceUnavailableException
 import re.notifica.NotificareNotReadyException
+import re.notifica.internal.NOTIFICARE_VERSION
 import re.notifica.internal.NotificareLogger
 import re.notifica.internal.NotificareModule
 import re.notifica.internal.NotificareUtils
 import re.notifica.internal.common.filterNotNull
-import re.notifica.internal.common.toByteArray
-import re.notifica.internal.common.toHex
 import re.notifica.internal.ktx.coroutineScope
 import re.notifica.internal.ktx.toCallbackFunction
+import re.notifica.internal.network.push.CreateDevicePayload
+import re.notifica.internal.network.push.CreateDeviceResponse
 import re.notifica.internal.network.push.DeviceDoNotDisturbResponse
-import re.notifica.internal.network.push.DeviceRegistrationPayload
 import re.notifica.internal.network.push.DeviceTagsPayload
 import re.notifica.internal.network.push.DeviceTagsResponse
 import re.notifica.internal.network.push.DeviceUpdateLanguagePayload
 import re.notifica.internal.network.push.DeviceUpdateTimeZonePayload
 import re.notifica.internal.network.push.DeviceUserDataResponse
 import re.notifica.internal.network.push.TestDeviceRegistrationPayload
+import re.notifica.internal.network.push.UpdateDeviceDoNotDisturbPayload
+import re.notifica.internal.network.push.UpdateDevicePayload
+import re.notifica.internal.network.push.UpdateDeviceUserDataPayload
+import re.notifica.internal.network.push.UpdateDeviceUserPayload
+import re.notifica.internal.network.push.UpgradeToLongLivedDevicePayload
 import re.notifica.internal.network.request.NotificareRequest
+import re.notifica.internal.storage.preferences.entities.StoredDevice
+import re.notifica.internal.storage.preferences.ktx.asPublic
 import re.notifica.ktx.eventsImplementation
 import re.notifica.ktx.session
 import re.notifica.models.NotificareDevice
 import re.notifica.models.NotificareDoNotDisturb
-import re.notifica.models.NotificareTransport
 import re.notifica.models.NotificareUserData
+import java.util.Locale
 
 @Keep
-internal object NotificareDeviceModuleImpl :
-    NotificareModule(),
-    NotificareDeviceModule,
-    NotificareInternalDeviceModule {
+internal object NotificareDeviceModuleImpl : NotificareModule(), NotificareDeviceModule {
+
+    private var storedDevice: StoredDevice?
+        get() = Notificare.sharedPreferences.device
+        set(value) {
+            Notificare.sharedPreferences.device = value
+        }
+
+    private var hasPendingDeviceRegistrationEvent: Boolean? = null
 
     // region Notificare Module
 
     override suspend fun launch() {
-        val device = currentDevice
+        upgradeToLongLivedDeviceWhenNeeded()
 
-        if (device != null) {
-            register(
-                transport = device.transport,
-                token = device.id,
-                userId = device.userId,
-                userName = device.userName,
-            )
+        val storedDevice = storedDevice
+
+        if (storedDevice == null) {
+            NotificareLogger.debug("New install detected")
+
+            createDevice()
+            hasPendingDeviceRegistrationEvent = true
 
             // Ensure a session exists for the current device.
             Notificare.session().launch()
 
-            if (device.appVersion != NotificareUtils.applicationVersion) {
+            // We will log the Install & Registration events here since this will execute only one time at the start.
+            Notificare.eventsImplementation().logApplicationInstall()
+            Notificare.eventsImplementation().logApplicationRegistration()
+        } else {
+            val isApplicationUpgrade = storedDevice.appVersion != NotificareUtils.applicationVersion
+
+            updateDevice()
+
+            // Ensure a session exists for the current device.
+            Notificare.session().launch()
+
+            if (isApplicationUpgrade) {
                 // It's not the same version, let's log it as an upgrade.
                 NotificareLogger.debug("New version detected")
                 Notificare.eventsImplementation().logApplicationUpgrade()
-            }
-        } else {
-            NotificareLogger.debug("New install detected")
-
-            try {
-                registerTemporary()
-
-                // Ensure a session exists for the current device.
-                Notificare.session().launch()
-
-                // We will log the Install & Registration events here since this will execute only one time at the start.
-                Notificare.eventsImplementation().logApplicationInstall()
-                Notificare.eventsImplementation().logApplicationRegistration()
-            } catch (e: Exception) {
-                NotificareLogger.warning("Failed to register temporary device.", e)
-                throw e
             }
         }
     }
 
     override suspend fun postLaunch() {
-        val device = currentDevice
-        if (device != null) notifyDeviceRegistered(device)
+        val device = storedDevice
+        if (device != null && hasPendingDeviceRegistrationEvent == true) {
+            notifyDeviceRegistered(device.asPublic())
+        }
     }
 
     // endregion
 
     // region Notificare Device Module
 
-    override var currentDevice: NotificareDevice?
-        get() = Notificare.sharedPreferences.device
-        private set(value) {
-            Notificare.sharedPreferences.device = value
-        }
+    override val currentDevice: NotificareDevice?
+        get() = Notificare.sharedPreferences.device?.asPublic()
 
     override val preferredLanguage: String?
         get() {
@@ -108,15 +109,49 @@ internal object NotificareDeviceModuleImpl :
             return "$preferredLanguage-$preferredRegion"
         }
 
+    @Deprecated(
+        message = "Use updateUser() instead.",
+        replaceWith = ReplaceWith("updateUser(userId, userName)"),
+    )
     override suspend fun register(userId: String?, userName: String?): Unit = withContext(Dispatchers.IO) {
-        checkPrerequisites()
-
-        val device = checkNotNull(currentDevice)
-        register(device.transport, device.id, userId, userName)
+        updateUser(
+            userId = userId,
+            userName = userName,
+        )
     }
 
+    @Deprecated(
+        message = "Use updateUser() instead.",
+        replaceWith = ReplaceWith("updateUser(userId, userName, callback)"),
+    )
+    @Suppress("DEPRECATION")
     override fun register(userId: String?, userName: String?, callback: NotificareCallback<Unit>): Unit =
         toCallbackFunction(::register)(userId, userName, callback)
+
+    override suspend fun updateUser(userId: String?, userName: String?): Unit = withContext(Dispatchers.IO) {
+        checkPrerequisites()
+
+        val device = storedDevice
+            ?: throw NotificareDeviceUnavailableException()
+
+        val payload = UpdateDeviceUserPayload(
+            userId = userId,
+            userName = userName,
+        )
+
+        NotificareRequest.Builder()
+            .put("/push/${device.id}", payload)
+            .response()
+
+        // Update current device properties.
+        storedDevice = device.copy(
+            userId = userId,
+            userName = userName,
+        )
+    }
+
+    override fun updateUser(userId: String?, userName: String?, callback: NotificareCallback<Unit>): Unit =
+        toCallbackFunction(::updateUser)(userId, userName, callback)
 
     override suspend fun updatePreferredLanguage(preferredLanguage: String?): Unit = withContext(Dispatchers.IO) {
         checkPrerequisites()
@@ -153,10 +188,10 @@ internal object NotificareDeviceModuleImpl :
     override suspend fun fetchTags(): List<String> = withContext(Dispatchers.IO) {
         checkPrerequisites()
 
-        val device = checkNotNull(currentDevice)
+        val device = checkNotNull(storedDevice)
 
         NotificareRequest.Builder()
-            .get("/device/${device.id}/tags")
+            .get("/push/${device.id}/tags")
             .responseDecodable(DeviceTagsResponse::class)
             .tags
     }
@@ -174,9 +209,9 @@ internal object NotificareDeviceModuleImpl :
     override suspend fun addTags(tags: List<String>): Unit = withContext(Dispatchers.IO) {
         checkPrerequisites()
 
-        val device = checkNotNull(currentDevice)
+        val device = checkNotNull(storedDevice)
         NotificareRequest.Builder()
-            .put("/device/${device.id}/addtags", DeviceTagsPayload(tags))
+            .put("/push/${device.id}/addtags", DeviceTagsPayload(tags))
             .response()
     }
 
@@ -193,9 +228,9 @@ internal object NotificareDeviceModuleImpl :
     override suspend fun removeTags(tags: List<String>): Unit = withContext(Dispatchers.IO) {
         checkPrerequisites()
 
-        val device = checkNotNull(currentDevice)
+        val device = checkNotNull(storedDevice)
         NotificareRequest.Builder()
-            .put("/device/${device.id}/removetags", DeviceTagsPayload(tags))
+            .put("/push/${device.id}/removetags", DeviceTagsPayload(tags))
             .response()
     }
 
@@ -205,9 +240,9 @@ internal object NotificareDeviceModuleImpl :
     override suspend fun clearTags(): Unit = withContext(Dispatchers.IO) {
         checkPrerequisites()
 
-        val device = checkNotNull(currentDevice)
+        val device = checkNotNull(storedDevice)
         NotificareRequest.Builder()
-            .put("/device/${device.id}/cleartags", null)
+            .put("/push/${device.id}/cleartags", null)
             .response()
     }
 
@@ -217,14 +252,14 @@ internal object NotificareDeviceModuleImpl :
     override suspend fun fetchDoNotDisturb(): NotificareDoNotDisturb? = withContext(Dispatchers.IO) {
         checkPrerequisites()
 
-        val device = checkNotNull(currentDevice)
+        val device = checkNotNull(storedDevice)
         val dnd = NotificareRequest.Builder()
-            .get("/device/${device.id}/dnd")
+            .get("/push/${device.id}/dnd")
             .responseDecodable(DeviceDoNotDisturbResponse::class)
             .dnd
 
         // Update current device properties.
-        currentDevice = device.copy(dnd = dnd)
+        storedDevice = device.copy(dnd = dnd)
 
         return@withContext dnd
     }
@@ -235,13 +270,18 @@ internal object NotificareDeviceModuleImpl :
     override suspend fun updateDoNotDisturb(dnd: NotificareDoNotDisturb): Unit = withContext(Dispatchers.IO) {
         checkPrerequisites()
 
-        val device = checkNotNull(currentDevice)
+        val device = checkNotNull(storedDevice)
+
+        val payload = UpdateDeviceDoNotDisturbPayload(
+            dnd = dnd,
+        )
+
         NotificareRequest.Builder()
-            .put("/device/${device.id}/dnd", dnd)
+            .put("/push/${device.id}", payload)
             .response()
 
         // Update current device properties.
-        currentDevice = device.copy(dnd = dnd)
+        storedDevice = device.copy(dnd = dnd)
     }
 
     override fun updateDoNotDisturb(dnd: NotificareDoNotDisturb, callback: NotificareCallback<Unit>): Unit =
@@ -250,13 +290,18 @@ internal object NotificareDeviceModuleImpl :
     override suspend fun clearDoNotDisturb(): Unit = withContext(Dispatchers.IO) {
         checkPrerequisites()
 
-        val device = checkNotNull(currentDevice)
+        val device = checkNotNull(storedDevice)
+
+        val payload = UpdateDeviceDoNotDisturbPayload(
+            dnd = null,
+        )
+
         NotificareRequest.Builder()
-            .put("/device/${device.id}/cleardnd", null)
+            .put("/push/${device.id}", payload)
             .response()
 
         // Update current device properties.
-        currentDevice = device.copy(dnd = null)
+        storedDevice = device.copy(dnd = null)
     }
 
     override fun clearDoNotDisturb(callback: NotificareCallback<Unit>): Unit =
@@ -265,15 +310,15 @@ internal object NotificareDeviceModuleImpl :
     override suspend fun fetchUserData(): NotificareUserData = withContext(Dispatchers.IO) {
         checkPrerequisites()
 
-        val device = checkNotNull(currentDevice)
+        val device = checkNotNull(storedDevice)
         val userData = NotificareRequest.Builder()
-            .get("/device/${device.id}/userdata")
+            .get("/push/${device.id}/userdata")
             .responseDecodable(DeviceUserDataResponse::class)
             .userData?.filterNotNull { it.value }
             ?: mapOf()
 
         // Update current device properties.
-        currentDevice = device.copy(userData = userData)
+        storedDevice = device.copy(userData = userData)
 
         return@withContext userData
     }
@@ -284,13 +329,18 @@ internal object NotificareDeviceModuleImpl :
     override suspend fun updateUserData(userData: NotificareUserData): Unit = withContext(Dispatchers.IO) {
         checkPrerequisites()
 
-        val device = checkNotNull(currentDevice)
+        val device = checkNotNull(storedDevice)
+
+        val payload = UpdateDeviceUserDataPayload(
+            userData = userData,
+        )
+
         NotificareRequest.Builder()
-            .put("/device/${device.id}/userdata", userData)
+            .put("/push/${device.id}", payload)
             .response()
 
         // Update current device properties.
-        currentDevice = device.copy(userData = userData)
+        storedDevice = device.copy(userData = userData)
     }
 
     override fun updateUserData(userData: NotificareUserData, callback: NotificareCallback<Unit>): Unit =
@@ -298,57 +348,20 @@ internal object NotificareDeviceModuleImpl :
 
     // endregion
 
-    // region Notificare Internal Device Module
-
-    override suspend fun registerTemporary(): Unit = withContext(Dispatchers.IO) {
-        val device = currentDevice
-
-        // NOTE: keep the same token if available and only when not changing transport providers.
-        val token = when {
-            device != null && device.transport == NotificareTransport.NOTIFICARE -> device.id
-            else -> UUID.randomUUID().toByteArray().toHex()
-        }
-
-        register(
-            transport = NotificareTransport.NOTIFICARE,
-            token = token,
-            userId = currentDevice?.userId,
-            userName = currentDevice?.userName,
-        )
-    }
-
-    override suspend fun registerPushToken(
-        transport: NotificareTransport,
-        token: String
-    ): Unit = withContext(Dispatchers.IO) {
-        if (transport == NotificareTransport.NOTIFICARE) {
-            throw IllegalArgumentException("Invalid transport '$transport'.")
-        }
-
-        register(
-            transport = transport,
-            token = token,
-            userId = currentDevice?.userId,
-            userName = currentDevice?.userName,
-        )
-    }
-
-    // endregion
-
     internal suspend fun delete(): Unit = withContext(Dispatchers.IO) {
         checkPrerequisites()
 
-        val device = checkNotNull(currentDevice)
+        val device = checkNotNull(storedDevice)
 
         NotificareRequest.Builder()
             .delete(
-                url = "/device/${device.id}",
+                url = "/push/${device.id}",
                 body = null,
             )
             .response()
 
         // Remove current device.
-        currentDevice = null
+        storedDevice = null
     }
 
     @Throws
@@ -359,74 +372,138 @@ internal object NotificareDeviceModuleImpl :
         }
     }
 
-    private suspend fun register(transport: NotificareTransport, token: String, userId: String?, userName: String?) {
-        if (registrationChanged(token, userId, userName)) {
-            val currentDevice = currentDevice
+    private suspend fun createDevice(): Unit = withContext(Dispatchers.IO) {
+        val payload = CreateDevicePayload(
+            language = getDeviceLanguage(),
+            region = getDeviceRegion(),
+            platform = "Android",
+            osVersion = NotificareUtils.osVersion,
+            sdkVersion = NOTIFICARE_VERSION,
+            appVersion = NotificareUtils.applicationVersion,
+            deviceString = NotificareUtils.deviceString,
+            timeZoneOffset = NotificareUtils.timeZoneOffset,
+            backgroundAppRefresh = true,
+        )
 
-            val oldDeviceId =
-                if (currentDevice?.id != null && currentDevice.id != token) currentDevice.id
-                else null
+        val response = NotificareRequest.Builder()
+            .post("/push", payload)
+            .responseDecodable(CreateDeviceResponse::class)
 
-            val deviceRegistration = DeviceRegistrationPayload(
-                deviceId = token,
-                oldDeviceId = oldDeviceId,
-                userId = userId,
-                userName = userName,
-                language = getDeviceLanguage(),
-                region = getDeviceRegion(),
-                platform = "Android",
-                transport = transport,
-                osVersion = NotificareUtils.osVersion,
-                sdkVersion = Notificare.SDK_VERSION,
-                appVersion = NotificareUtils.applicationVersion,
-                deviceString = NotificareUtils.deviceString,
-                timeZoneOffset = NotificareUtils.timeZoneOffset,
-                backgroundAppRefresh = true,
-
-                // Submit a value when registering a temporary to prevent
-                // otherwise let the push module take over and update the setting accordingly.
-                allowedUI = if (transport == NotificareTransport.NOTIFICARE) false else null
-            )
-
-            NotificareRequest.Builder()
-                .post("/device", deviceRegistration)
-                .response()
-
-            deviceRegistration.toStoredDevice(currentDevice).also {
-                this.currentDevice = it
-            }
-        } else {
-            NotificareLogger.info("Skipping device registration, nothing changed.")
-        }
-
-        if (Notificare.isReady) {
-            val device = checkNotNull(currentDevice)
-            notifyDeviceRegistered(device)
-        }
+        storedDevice = StoredDevice(
+            id = response.device.deviceId,
+            userId = null,
+            userName = null,
+            timeZoneOffset = payload.timeZoneOffset,
+            osVersion = payload.osVersion,
+            sdkVersion = payload.sdkVersion,
+            appVersion = payload.appVersion,
+            deviceString = payload.deviceString,
+            language = payload.language,
+            region = payload.region,
+            dnd = null,
+            userData = mapOf(),
+        )
     }
 
-    private fun registrationChanged(token: String?, userId: String?, userName: String?): Boolean {
-        val device = currentDevice ?: run {
+    private suspend fun updateDevice(): Unit = withContext(Dispatchers.IO) {
+        val storedDevice = checkNotNull(storedDevice)
+
+        if (!registrationChanged()) {
+            NotificareLogger.debug("Skipping device update, nothing changed.")
+            return@withContext
+        }
+
+        val payload = UpdateDevicePayload(
+            language = getDeviceLanguage(),
+            region = getDeviceRegion(),
+            platform = "Android",
+            osVersion = NotificareUtils.osVersion,
+            sdkVersion = NOTIFICARE_VERSION,
+            appVersion = NotificareUtils.applicationVersion,
+            deviceString = NotificareUtils.deviceString,
+            timeZoneOffset = NotificareUtils.timeZoneOffset,
+        )
+
+        NotificareRequest.Builder()
+            .put("/push/${storedDevice.id}", payload)
+            .response()
+
+        this@NotificareDeviceModuleImpl.storedDevice = storedDevice.copy(
+            timeZoneOffset = payload.timeZoneOffset,
+            osVersion = payload.osVersion,
+            sdkVersion = payload.sdkVersion,
+            appVersion = payload.appVersion,
+            deviceString = payload.deviceString,
+            language = payload.language,
+            region = payload.region,
+            dnd = storedDevice.dnd,
+            userData = storedDevice.userData,
+        )
+    }
+
+    private suspend fun upgradeToLongLivedDeviceWhenNeeded(): Unit = withContext(Dispatchers.IO) {
+        val currentDevice = Notificare.sharedPreferences.device
+            ?: return@withContext
+
+        if (currentDevice.isLongLived) return@withContext
+
+        NotificareLogger.info("Upgrading current device from legacy format.")
+
+        val deviceId = currentDevice.id
+        val transport = checkNotNull(currentDevice.transport)
+        val subscriptionId = if (transport != "Notificare") deviceId else null
+
+        val payload = UpgradeToLongLivedDevicePayload(
+            deviceId = deviceId,
+            transport = transport,
+            subscriptionId = subscriptionId,
+            language = currentDevice.language,
+            region = currentDevice.region,
+            platform = "Android",
+            osVersion = currentDevice.osVersion,
+            sdkVersion = currentDevice.sdkVersion,
+            appVersion = currentDevice.appVersion,
+            deviceString = currentDevice.deviceString,
+            timeZoneOffset = currentDevice.timeZoneOffset,
+        )
+
+        @Suppress("MagicNumber")
+        val response = NotificareRequest.Builder()
+            .post("/push", payload)
+            .responseDecodable { responseCode ->
+                when (responseCode) {
+                    201 -> CreateDeviceResponse::class
+                    else -> null
+                }
+            }
+
+        if (response != null) {
+            NotificareLogger.debug("New device identifier created.")
+        }
+
+        storedDevice = StoredDevice(
+            id = response?.device?.deviceId ?: currentDevice.id,
+            userId = currentDevice.userId,
+            userName = currentDevice.userName,
+            timeZoneOffset = currentDevice.timeZoneOffset,
+            osVersion = currentDevice.osVersion,
+            sdkVersion = currentDevice.sdkVersion,
+            appVersion = currentDevice.appVersion,
+            deviceString = currentDevice.deviceString,
+            language = currentDevice.language,
+            region = currentDevice.region,
+            dnd = currentDevice.dnd,
+            userData = currentDevice.userData,
+        )
+    }
+
+    private fun registrationChanged(): Boolean {
+        val device = storedDevice ?: run {
             NotificareLogger.debug("Registration check: fresh installation")
             return true
         }
 
         var changed = false
-
-        if (device.userId != userId) {
-            NotificareLogger.debug("Registration check: user id changed")
-            changed = true
-        }
-
-        if (device.userName != userName) {
-            NotificareLogger.debug("Registration check: user name changed")
-            changed = true
-        }
-
-        if (device.id != token) {
-            NotificareLogger.debug("Registration check: device token changed")
-            changed = true
-        }
 
         if (device.deviceString != NotificareUtils.deviceString) {
             NotificareLogger.debug("Registration check: device string changed")
@@ -463,16 +540,6 @@ internal object NotificareDeviceModuleImpl :
             changed = true
         }
 
-        val oneDayAgo = GregorianCalendar().apply { add(Calendar.DAY_OF_YEAR, -1) }.time
-        if (device.lastRegistered.before(oneDayAgo)) {
-            NotificareLogger.debug("Registration check: region changed")
-            changed = true
-        }
-
-        // TODO check the properties below
-        // v2 Android checks for:
-        // transport, allowedUI, locationServicesAuthStatus, bluetoothEnabled
-
         return changed
     }
 
@@ -492,7 +559,7 @@ internal object NotificareDeviceModuleImpl :
                         .put(
                             url = "/support/testdevice/$nonce",
                             body = TestDeviceRegistrationPayload(
-                                deviceId = checkNotNull(currentDevice).id,
+                                deviceId = checkNotNull(storedDevice).id,
                             ),
                         )
                         .response()
@@ -508,11 +575,11 @@ internal object NotificareDeviceModuleImpl :
     internal suspend fun updateLanguage(language: String, region: String) {
         checkPrerequisites()
 
-        val device = checkNotNull(currentDevice)
+        val device = checkNotNull(storedDevice)
 
         NotificareRequest.Builder()
             .put(
-                url = "/device/${device.id}",
+                url = "/push/${device.id}",
                 body = DeviceUpdateLanguagePayload(
                     language = language,
                     region = region,
@@ -521,7 +588,7 @@ internal object NotificareDeviceModuleImpl :
             .response()
 
         // Update current device properties.
-        currentDevice = device.copy(
+        storedDevice = device.copy(
             language = language,
             region = region,
         )
@@ -530,11 +597,11 @@ internal object NotificareDeviceModuleImpl :
     internal suspend fun updateTimeZone() {
         checkPrerequisites()
 
-        val device = checkNotNull(currentDevice)
+        val device = checkNotNull(storedDevice)
 
         NotificareRequest.Builder()
             .put(
-                url = "/device/${device.id}",
+                url = "/push/${device.id}",
                 body = DeviceUpdateTimeZonePayload(
                     language = getDeviceLanguage(),
                     region = getDeviceRegion(),
@@ -551,23 +618,4 @@ internal object NotificareDeviceModuleImpl :
                 .putExtra(Notificare.INTENT_EXTRA_DEVICE, device)
         )
     }
-}
-
-private fun DeviceRegistrationPayload.toStoredDevice(previous: NotificareDevice?): NotificareDevice {
-    return NotificareDevice(
-        id = this.deviceId,
-        userId = this.userId,
-        userName = this.userName,
-        timeZoneOffset = this.timeZoneOffset,
-        osVersion = this.osVersion,
-        sdkVersion = this.sdkVersion,
-        appVersion = this.appVersion,
-        deviceString = this.deviceString,
-        language = this.language,
-        region = this.region,
-        transport = this.transport,
-        dnd = previous?.dnd,
-        userData = previous?.userData ?: mapOf(),
-        lastRegistered = Date(),
-    )
 }